libnode: move GetSupportedCPUModels to package
