+++ tests/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "//pkg/certificates/triple/cert:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/util:go_default_library",
-        "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-operator/resource/generate/components:go_default_library",
         "//pkg/virt-operator/util:go_default_library",
